Using direct database handling instead of JComponentHelper to fetch component parameters
Under some circumstances, the component may have to modify the data stored in the component parameters before Live Update loads. This will cause an inconsistency between the cached data served by JComponentHelper and the data in the database. Having Live Update store its update status under these circumstances will cause the data modifications/additions to be lost. With this workaround, this doesn't happen as long as the component uses the same trick.